+++ src/views/Transfer/Swap/Swap.send.vue
@@ -32,10 +32,11 @@
 
   <nomad-button
     v-if="checkingLiquidity"
-    class="w-full uppercase mt-6 bg-white text-black h-11 flex flex-row justify-center opacity-70 cursor-default"
+    class="w-full uppercase mt-6 bg-white text-black h-11 flex flex-row justify-center bg-opacity-70 cursor-default"
   >
-    Checking availability
-    <loader-bounce color="black" class="ml-1" />
+    Checking
+    <img src="@/assets/connext-logo-black.png" class="h-6 px-2" />
+    availability
   </nomad-button>
 
   <nomad-button
@@ -62,7 +63,6 @@ import { NText, NDivider, useNotification } from 'naive-ui'
 import { useStore } from '@/store'
 import NomadButton from '@/components/Button.vue'
 import BridgeQuote from './Swap.quote.vue'
-import LoaderBounce from '@/components/LoaderBounce.vue'
 
 export default defineComponent({
   components: {
@@ -70,7 +70,6 @@ export default defineComponent({
     NDivider,
     NomadButton,
     BridgeQuote,
-    LoaderBounce,
   },
   props: {
     v$: {