+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This change log is manually updated at the moment.
 
 Unreleased
 --------------------------------
+- Update copy w/logo, "Checking Connext availability"
 
 v1.0.2 (Jan 14th, 2022)
 --------------------------------